refactor: 3주차 과제 진행 전 리팩토링 및 2주차 과제 피드백 반영

+++ apps/commerce-api/src/main/java/com/loopers/domain/point/PointHistory.java
@@ -8,24 +8,26 @@ import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @NoArgsConstructor(access = lombok.AccessLevel.PROTECTED)
 @Entity
 @Table(name = "point_histories")
 public class PointHistory extends BaseEntity {
     private Long userPk;
-    private Long point;
+    private BigDecimal point;
 
     @Enumerated(EnumType.STRING)
     private PointHistoryType type;
 
-    private PointHistory(Long userPk, Long point, PointHistoryType pointHistoryType) {
+    private PointHistory(Long userPk, BigDecimal point, PointHistoryType pointHistoryType) {
         this.userPk = userPk;
         this.point = point;
         this.type = pointHistoryType;
     }
 
-    public static PointHistory create(Long userPk, Long point, PointHistoryType pointHistoryType) {
+    public static PointHistory create(Long userPk, BigDecimal point, PointHistoryType pointHistoryType) {
         return new PointHistory(userPk, point, pointHistoryType);
     }
